+++ docs/source/conf.py
@@ -21,13 +21,13 @@ print(os.path.abspath('../../'))
 # -- Project information -----------------------------------------------------
 
 project = 'LexNLP'
-copyright = '2018, ContraxSuite, LLC'
+copyright = '2017-2019, ContraxSuite, LLC'
 author = 'ContraxSuite, LLC'
 
 # The short X.Y version
-version = '0.2.5'
+version = '0.2.6'
 # The full version, including alpha/beta/rc tags
-release = '0.2.5'
+release = '0.2.6'
 
 # -- General configuration ---------------------------------------------------