+++ lexnlp/extract/en/entities/nltk_maxent.py
@@ -26,7 +26,7 @@ from lexnlp.nlp.en.tokens import get_token_list
 __author__ = "ContraxSuite, LLC; LexPredict, LLC"
 __copyright__ = "Copyright 2015-2019, ContraxSuite, LLC"
 __license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/master/LICENSE"
-__version__ = "0.2.5"
+__version__ = "0.2.6"
 __maintainer__ = "LexPredict, LLC"
 __email__ = "support@contraxsuite.com"
 
@@ -239,7 +239,6 @@ def get_noun_phrases(text, strict=False, return_source=False, window=3, valid_pu
                 yield nnp, sentence
             else:
                 yield nnp
-
 
 
 class CompanyNPExtractor(NPExtractor):