+++ lexnlp/extract/en/tests/test_geoentities.py
@@ -5,18 +5,17 @@
 This module implements unit tests for the geo entity extraction functionality in English.
 
 """
-import csv
 import os
 
-from lexnlp.extract.en.dict_entities import get_entity_name, entity_config, add_aliases_to_entity, \
+from lexnlp.extract.en.dict_entities import get_entity_name, \
     prepare_alias_blacklist_dict
-from lexnlp.extract.en.geoentities import get_geoentities
+from lexnlp.extract.en.geoentities import get_geoentities, load_entities_dict_by_path
 from lexnlp.tests import lexnlp_tests
 
 __author__ = "ContraxSuite, LLC; LexPredict, LLC"
 __copyright__ = "Copyright 2015-2019, ContraxSuite, LLC"
 __license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/master/LICENSE"
-__version__ = "0.2.5"
+__version__ = "0.2.6"
 __maintainer__ = "LexPredict, LLC"
 __email__ = "support@contraxsuite.com"
 
@@ -24,25 +23,7 @@ __email__ = "support@contraxsuite.com"
 def load_entities_dict():
     entities_fn = os.path.join(os.path.dirname(lexnlp_tests.this_test_data_path()), 'geoentities.csv')
     aliases_fn = os.path.join(os.path.dirname(lexnlp_tests.this_test_data_path()), 'geoaliases.csv')
-
-    entities = {}
-
-    with open(entities_fn, 'r', encoding='utf8') as f:
-        reader = csv.DictReader(f)
-        for row in reader:
-            entities[row['id']] = entity_config(row['id'], row['name'], int(row['priority']) if row['priority'] else 0,
-                                                name_is_alias=True)
-
-    with open(aliases_fn, 'r', encoding='utf8') as f:
-        reader = csv.DictReader(f)
-        for row in reader:
-            entity = entities.get(row['entity_id'])
-            if entity:
-                add_aliases_to_entity(entity,
-                                      row['alias'],
-                                      row['locale'],
-                                      row['type'].startswith('iso') or row['type'] == 'abbreviation')
-    return entities.values()
+    return load_entities_dict_by_path(entities_fn, aliases_fn)
 
 
 _CONFIG = list(load_entities_dict())