+++ scripts/create_release_branch.sh
@@ -1,21 +1,21 @@
 #!/usr/bin/env bash
 
 # ARGS:
-# 1. '-core' repository DEVELOP branch name
+# 1. '-core' repository DEV RELEASE branch name
 # 2. '-lexnlp' repository OLD release branch name
 # 3. '-lexnlp' repository NEW release branch name
 
 CORE_REPO_PATH="/home/alex/dev/michael/contraxsuite/lexpredict-contraxsuite-core"
 LEXNLP_REPO_PATH="/home/alex/dev/michael/contraxsuite/lexpredict-lexnlp"
-DEV_BRANCH=$1
+DEV_RELEASE_BRANCH=$1
 OLD_RELEASE_BRANCH=$2
 NEW_RELEASE_BRANCH=$3
-aline="================================================================="
+LINE="================================================================="
 
 # prompt
 if [ $# -ne 3 ]; then
-    read -p "Enter '-core' repository DEVELOP branch name: " DEV_BRANCH
-    if [ -z ${DEV_BRANCH} ]; then
+    read -p "Enter '-core' repository DEVELOP branch name: " DEV_RELEASE_BRANCH
+    if [ -z ${DEV_RELEASE_BRANCH} ]; then
         echo "Exiting program."
         return 1
     fi
@@ -32,8 +32,8 @@ if [ $# -ne 3 ]; then
 fi
 
 # confirm further processing
-echo ${aline}
-echo "Copy data from core@${DEV_BRANCH} to lexnlp@${NEW_RELEASE_BRANCH}, continue?" yn
+echo ${LINE}
+echo "Copy data from core@${DEV_RELEASE_BRANCH} to lexnlp@${NEW_RELEASE_BRANCH}, continue?" yn
 select yn in "Yes" "No"; do
     case ${yn} in
         Yes ) break;;
@@ -42,36 +42,44 @@ select yn in "Yes" "No"; do
 done
 
 # check if develop branch exists
-echo ${aline}
-echo "Update core@${DEV_BRANCH}"
+echo ${LINE}
+echo "Update core@${DEV_RELEASE_BRANCH}"
 
 pushd ${CORE_REPO_PATH}
 
-if [ -z "`git branch --list ${DEV_BRANCH}`" ]
+if [ -z "`git branch --list ${DEV_RELEASE_BRANCH}`" ]
 then
-   echo "'-core' branch ${DEV_BRANCH} doesn't exist. Exiting program."
+   echo "'-core' branch ${DEV_RELEASE_BRANCH} doesn't exist. Exiting program."
    return 1
 fi
-git pull origin ${DEV_BRANCH}
+git pull origin ${DEV_RELEASE_BRANCH}
 
 # replace old release number with new one
-echo ${aline}
-echo "Update version number in core@${DEV_BRANCH} from ${OLD_RELEASE_BRANCH} to ${NEW_RELEASE_BRANCH}"
+echo ${LINE}
+echo "Update version number in core@${DEV_RELEASE_BRANCH} from ${OLD_RELEASE_BRANCH} to ${NEW_RELEASE_BRANCH}"
 
 OLD_RELEASE_BRANCH_esc=$(echo ${OLD_RELEASE_BRANCH} | sed 's,\.,\\.,g')
 NEW_RELEASE_BRANCH_esc=$(echo ${NEW_RELEASE_BRANCH} | sed 's,\.,\\.,g')
 find ./ -type f -readable -writable -exec sed -i "s/__version__ = \"$OLD_RELEASE_BRANCH_esc\"/__version__ = \"$NEW_RELEASE_BRANCH_esc\"/g" {} \;
 find setup.py -type f -readable -writable -exec sed -i "s/version='$OLD_RELEASE_BRANCH_esc'/version='$NEW_RELEASE_BRANCH_esc'/g" {} \;
+find docs/source/conf.py -type f -readable -writable -exec sed -i "s/version = '$OLD_RELEASE_BRANCH_esc'/version = '$NEW_RELEASE_BRANCH_esc'/g" {} \;
+find docs/source/conf.py -type f -readable -writable -exec sed -i "s/release = '$OLD_RELEASE_BRANCH_esc'/release = '$NEW_RELEASE_BRANCH_esc'/g" {} \;
 
-echo ${aline}
-echo "Update version number in core@${DEV_BRANCH} from ${OLD_RELEASE_BRANCH} to ${NEW_RELEASE_BRANCH}"
+echo ${LINE}
+echo "Update version number in core@${DEV_RELEASE_BRANCH} from ${OLD_RELEASE_BRANCH} to ${NEW_RELEASE_BRANCH}"
 git add --all
 git commit -m "CS: updated version from ${OLD_RELEASE_BRANCH} to ${NEW_RELEASE_BRANCH}"
-git push origin ${DEV_BRANCH}
+git push origin ${DEV_RELEASE_BRANCH}
+
+# tag latest release commit
+echo ${LINE}
+echo "Tag latest release commit in private@${DEV_RELEASE_BRANCH} branch as ${NEW_RELEASE_BRANCH}"
+#git tag ${NEW_RELEASE_BRANCH}
+#git push origin --tags
 
 popd
 
-echo ${aline}
+echo ${LINE}
 echo "Update lexnlp@${OLD_RELEASE_BRANCH}"
 
 pushd ${LEXNLP_REPO_PATH}
@@ -87,7 +95,7 @@ git fetch origin ${OLD_RELEASE_BRANCH}
 git checkout ${OLD_RELEASE_BRANCH}
 
 # check that new release branch exists, checkout/create
-echo ${aline}
+echo ${LINE}
 echo "Create lexnlp@${NEW_RELEASE_BRANCH}"
 
 if [ -z "`git branch --list ${NEW_RELEASE_BRANCH}`" ]
@@ -98,31 +106,39 @@ else
    git checkout ${NEW_RELEASE_BRANCH}
 fi
 
+
+find ${CORE_REPO_PATH}/ -name "*pyc" -delete
+
 # copy files from develop to new branch
-echo ${aline}
+echo ${LINE}
 echo "Copy files from -core to -lexnlp local repo"
 
-cp -rf ${CORE_REPO_PATH}/lexnlp ${LEXNLP_REPO_PATH}
-cp -rf ${CORE_REPO_PATH}/libs ${LEXNLP_REPO_PATH}
-cp -rf ${CORE_REPO_PATH}/scripts ${LEXNLP_REPO_PATH}
-cp -rf ${CORE_REPO_PATH}/test_data ${LEXNLP_REPO_PATH}
+rsync -av --delete ${CORE_REPO_PATH}/lexnlp/ ${LEXNLP_REPO_PATH}/lexnlp
+rsync -av --delete ${CORE_REPO_PATH}/libs/ ${LEXNLP_REPO_PATH}/libs
+rsync -av --delete ${CORE_REPO_PATH}/docs/ ${LEXNLP_REPO_PATH}/docs
+rsync -av --delete ${CORE_REPO_PATH}/scripts/ ${LEXNLP_REPO_PATH}/scripts
+rsync -av --delete ${CORE_REPO_PATH}/test_data/ ${LEXNLP_REPO_PATH}/test_data
+rsync -av --delete ${CORE_REPO_PATH}/lexnlp/ ${LEXNLP_REPO_PATH}/lexnlp
 cp -rf ${CORE_REPO_PATH}/python-requirements* ${LEXNLP_REPO_PATH}
 cp -rf ${CORE_REPO_PATH}/setup.py ${LEXNLP_REPO_PATH}
 cp -rf ${CORE_REPO_PATH}/MANIFEST.in ${LEXNLP_REPO_PATH}
+cp -rf ${CORE_REPO_PATH}/index.rst ${LEXNLP_REPO_PATH}
+cp -rf ${CORE_REPO_PATH}/README.rst ${LEXNLP_REPO_PATH}
+cp -rf ${CORE_REPO_PATH}/README.md ${LEXNLP_REPO_PATH}
 
 # create commit and push
-echo ${aline}
+echo ${LINE}
 echo "Commit and push new branch to -lexnlp remote repo"
-git add --all
-git commit -m "initial ${NEW_RELEASE_BRANCH} commit"
-git push origin ${NEW_RELEASE_BRANCH}
+#git add --all
+#git commit -m "initial ${NEW_RELEASE_BRANCH} commit"
+#git push origin ${NEW_RELEASE_BRANCH}
 
 # update master
-echo ${aline}
+echo ${LINE}
 echo "Update master branch in -lexnlp remote repo"
-git checkout master
-git pull origin master
-git merge ${NEW_RELEASE_BRANCH}
-git push origin master
+#git checkout master
+#git pull origin master
+#git merge ${NEW_RELEASE_BRANCH}
+#git push origin master
 
 popd